backend/requirements.txt
# Backend dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
aiofiles>=23.2.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
websockets>=12.0

# ML/Vision
torch>=2.1.0
transformers>=4.36.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0

# Video processing
numpy>=1.24.0
